Ensure govvv is populating the correct pkg with ldflags.
When using the Makefile to build Sherpa locally govvv was
attempting to populate the main package with the ldflags which was
not working as the build information needs to be populated into
the build package. The Makefile has therefore been updated and the
version formatting change slightly to ensure alignment.