Configure basic TLS server and client support.
This change allows the operator to configure the Sherpa server to
use TLS for traffic encryption and security. The TLS cert and key
are configurable via CLI flags or env vars as with all other
params.

To allow the CLI to communicate over HTTPS, the API config and
client have also been updated to accept TLS config params.