Update server CLI flags to be more consistent and user friendly.
Previosuly the CLI flags for the sherpa server were not very
consistent and did not follow any logic. This update groups the
CLI flags more appropriately, making them easier to work with and
understand. This refactor will also be helpful in the future when
writing the code to handle configuring a server using a config
file.