Add initial support for statsite and statsd runtime telemetry.
This adds support for configuring the Sherpa server to send metrics
to both statsite and statsd endpoints. The metrics currently sent
are only runtime metrics, but this will be expanded upon in due
course.

The docs have been updated to reflect the new CLI flags and a new
file is now available which details the metrics which are
available.