Add metrics API endpoint and command to retrive server telemetry.
In addition to sending telemetry to an endpoint, this commit now
allows metrics to be viewed by calling the Sherpa API. A command
can also be used, which wraps the API, to output the telemetry
information to the terminal. The API and command are nested under
the system commands and endpoints.