Internal autoscaling thesholds are now configurable within policy.
Previously the internal autoscaling thresholds for CPU and memory
were fixed at 80% and 20% using constants which could not be
changed by the user. This update now makes the thresholds
configurable via scaling policies so that users can customize the
thesholds per task group. If the user does not specify thresholds,
the 80:20 defaults will be used.