Fix bug here meta policy engine would not process job updates.
The meta policy watcher was doing an incorrect comparison when
checking whether the modify index on a job specification returned
from the API. Previously it was comparing against the maxfound
which was being set to the Jobs().List() returned meta. This meant
the job modify index would always be <= as it could never be
greater than the value of the API returned meta index.

In addition to fixing the above mentioned bug; this commit adds
some contextual logging around the meta watcher to help identify
any future issues and make operating in this mode easier.

Closes #15