Use goroutine pool within autoscaler to limit concurrency.
Previously the internal autoscaler iterated through the scaling
policies and launched a go routine to run the autoscaling check
process. There were no limits on the number of routines which
could lead to overloading on the Nomad API.

This change introduces a goroutine pool in order to limit the
number of concurrent scaling threads. The number of threads is
configurable via the CLI, with a default of 3. When all threads
are in use, the iteration will block until additional work
can be consumed. There is a timelimit of the execution, currently
set to 60's which we may want to make configurable in the future.