Add scaling event state management to track event details.
Previously the only way to understand what scaling events had been
triggered was to check the Sherpa server logs. This is not ideal
for operators and results in an incomplete picture. This change
adds scaling event state storage, allowing for operators to list
and view scaling events that have been triggered.

The Consul path CLI parameter has been changed to reflect the fact
it now relates to both policies and state. The change doesn't break
existing installs though.

Closes #26