Add simple UI to provide visual overview of scaling events.
It is sometimes easier for operators to get a quick overview of
system information by glancing at a UI. This commit therefore
introduces a UI which can be enabled via a CLI flag.

The UI provides an overview of scaling events in a table format
headed by links to useful Sherpa resoruces. When the UI is enabled
a call to / will be redirected to /ui where the UI endpoint handler
is running.

Closes #31