Autoscaler log to debug when it suggests a group should be scaled.
When the autoscaler determines a group can be scaled based on its
resource consumption the log line should be at debug rather than
info. This is because the scaler then needs to process the request
which may not be allowed, meaning this addition is advisory and
can clutter logs when a group is running on very low resource use.