Move Nomad Meta policy to satisfy policy interface.
Previously the Nomad meta policy engine conflicted with the Consul
storage backend. This was due to the early design of storage being
for policies only. Recently the storage backend has improved to
include additional state items, and therefore is not just for
policies. The Nomad meta policy engine should therefore be able to
run alongside the Consul state backend, with the Nomad meta
policies storage in memory. This is because we treat Nomad as the
source of truth.

The Nomad meta watcher is now also moved over to use the watcher
package, in the same way deployments are tracked. Several
improvements to the meta watcher are also included, such as
handling groups which have their policy removed, but the job
continues to run.