Add acceptance test for meta policy engine and scaling actions.
The acceptance testing has been expanded to cover scaling actions
as well as the Nomad meta policy backend. The Nomad meta engine
requires Sherpa to run in a different mode than the rest of the
tests, therefore the make targets have been split to make this
clear until the whole process is automated.

The scale tests are particulary basic, but are a start for further
work in the future.

These tests have already uncovered #59.