Fix incorrect response code and msg when scaling breaks thresholds
Previously errors when triggering a scaling action such as the
new count breaking policy thesholds would result in a 304 response
code and an empty message. This changes fixes the behaviour so
that the client will recieve a 409 response code with a useful
message to indicate the reason for not performing the action.

Closes #59