Add cooldown functionality to scaling policies and storage backend. (#77)
Sherpa policies now allow the specifying of a cooldown as a time
duration in seconds. When a scaling activity is requested, the
cooldown is checked against the last scaling event for the job
group. If the last event has occurred within the cooldown period,
the scaling request will be denied.

Closes #49