autoscale: provide additional decision context on scaling choices
When the autoscaler performs its logic to figure out whether a
scaling event should be triggered, it uses a case statement to
check if thresholds are broken. The problem with this is that the
first match is used, which is correct, but doesn't provide a great
amount of detail. It is possible that both out and in actions are
desired by different resources, or that multiple resources both
require the same action.

This change allows for greater insight into the scaling decision
made by the autoscaler calculation. It helps operators understand
if changes need to be made to group resource settings, or exactly
which thresholds were broken.