autoscale: meta submission allows multiple resource entrie.
Previously the way the meta was built within the autoscaler only
meant a single metric could be tracked properly. This causes
misinformation as it is possible a scaling activity happens due
to greater than one resource breaking its threshold. This change
therefore updates the way the autoscaling meta is formed, to allow
for multiple resources to be tracked.