server: add prometheus format option to metrics endpoint.
Previously the metrics endpoint exposed metrics in a format which
was not consumable by prometheus. This update allows users to
configure the telemetry setup to expose metrics in a prometheus
friendly manner. In metrics endpoint now also accepts a format
query param for use by Prometheus when scrapping to ensure metrics
are in the desired format.